move model processors to helpers directory
