+++ dashboard/app/helpers/arithmo_processor.rb
