+++ dashboard/app/helpers/karen_processor.rb
