+++ dashboard/app/helpers/mistral_processor.rb
