+++ dashboard/app/helpers/pirate_processor.rb
